save pomo details
